+++ bin/gen/app/CommonGen.js
@@ -1,14 +1,22 @@
 "use strict";
+var path = require("path");
 var ProjectGen_1 = require("../ProjectGen");
 var Util_1 = require("../../util/Util");
 var GitGen_1 = require("../file/GitGen");
 var Vesta_1 = require("../file/Vesta");
-var resolve = Promise.resolve;
 var CommonGen = (function () {
     function CommonGen(config) {
         this.config = config;
         this.vesta = Vesta_1.Vesta.getInstance();
     }
+    CommonGen.prototype.addSubModule = function () {
+        if (!this.config.repository.common)
+            return;
+        var dir = this.config.name;
+        var destDir = this.config.type == ProjectGen_1.ProjectGen.Type.ClientSide ? 'src/app/cmn' : 'src/cmn';
+        var repo = this.config.repository;
+        return Util_1.Util.execSync("git submodule add -b dev " + GitGen_1.GitGen.getRepoUrl(repo.baseUrl, repo.group, repo.common) + " " + destDir, dir);
+    };
     /**
      * git init
      * git commit --message=""
@@ -17,36 +25,23 @@ var CommonGen = (function () {
      * git checkout -b dev
      * git push origin dev
      */
-    CommonGen.prototype.initFirstTimeCommonProject = function () {
-        var repository = this.config.repository, projectRepo = this.vesta.getProjectConfig().repository, cmnDir = repository.common;
-        return GitGen_1.GitGen.getRepoUrl(projectRepo.baseRepoUrl)
-            .then(function (url) {
-            return GitGen_1.GitGen.clone(url + "/" + projectRepo.group + "/" + projectRepo.common + ".git", cmnDir);
-        })
-            .then(function () { return GitGen_1.GitGen.cleanClonedRepo(cmnDir); })
-            .then(function () { return Util_1.Util.exec("git init", cmnDir); })
-            .then(function () { return Util_1.Util.exec("git add .", cmnDir); })
-            .then(function () { return Util_1.Util.exec("git commit -m Vesta", cmnDir); })
-            .then(function () { return GitGen_1.GitGen.getRepoUrl(repository.baseRepoUrl, true); })
-            .then(function (sshUrl) { return Util_1.Util.exec("git remote add origin " + sshUrl + ":" + repository.group + "/" + repository.common + ".git", cmnDir); })
-            .then(function () { return Util_1.Util.exec("git push -u origin master", cmnDir); })
-            .then(function () { return Util_1.Util.exec("git checkout -b dev", cmnDir); })
-            .then(function () { return Util_1.Util.exec("git push -u origin dev", cmnDir); });
-    };
-    CommonGen.prototype.addSubModule = function () {
-        var _this = this;
-        if (!this.config.repository.common)
-            return resolve();
-        var dir = this.config.name, projectRepo = this.vesta.getProjectConfig().repository, destDir = this.config.type == ProjectGen_1.ProjectGen.Type.ClientSide ? 'src/app/cmn' : 'src/cmn';
-        return GitGen_1.GitGen.getRepoUrl(projectRepo.baseRepoUrl, true)
-            .then(function (url) { return Util_1.Util.exec("git submodule add -b dev " + url + ":" + _this.config.repository.group + "/" + _this.config.repository.common + ".git " + destDir, dir); });
+    CommonGen.prototype.createCommonProject = function () {
+        var repository = this.config.repository, templateRepo = this.vesta.getProjectConfig().repository, cmnDir = repository.common;
+        GitGen_1.GitGen.clone(GitGen_1.GitGen.getRepoUrl(templateRepo.baseUrl, templateRepo.group, templateRepo.common), cmnDir);
+        GitGen_1.GitGen.cleanClonedRepo(cmnDir);
+        Util_1.Util.execSync("git init", cmnDir);
+        Util_1.Util.execSync("git add .", cmnDir);
+        Util_1.Util.execSync("git commit -m Vesta-init", cmnDir);
+        Util_1.Util.execSync("git remote add origin " + GitGen_1.GitGen.getRepoUrl(repository.baseUrl, repository.group, repository.common), cmnDir);
+        Util_1.Util.execSync("git push -u origin master", cmnDir);
+        Util_1.Util.execSync("git checkout -b dev", cmnDir);
+        Util_1.Util.execSync("git push -u origin dev", cmnDir);
     };
     CommonGen.prototype.initWithoutSubModule = function () {
-        var dir = this.config.name, projectRepo = this.vesta.getProjectConfig().repository, destDir = this.config.type == ProjectGen_1.ProjectGen.Type.ClientSide ? 'src/app/cmn' : 'src/cmn';
-        Util_1.Util.fs.mkdir(destDir, dir + "/fw/common");
-        return GitGen_1.GitGen.getRepoUrl(projectRepo.baseRepoUrl)
-            .then(function (url) { return GitGen_1.GitGen.clone(url + "/" + projectRepo.group + "/" + projectRepo.common + ".git", dir + "/fw/common"); })
-            .then(function () { return Util_1.Util.fs.copy(dir + "/fw/common", destDir); });
+        var dir = this.config.name, templateRepo = this.vesta.getProjectConfig().repository, destDir = path.join(dir, this.config.type == ProjectGen_1.ProjectGen.Type.ClientSide ? 'src/app/cmn' : 'src/cmn');
+        Util_1.Util.fs.mkdir(destDir);
+        GitGen_1.GitGen.clone(GitGen_1.GitGen.getRepoUrl(templateRepo.baseUrl, templateRepo.group, templateRepo.common), destDir);
+        GitGen_1.GitGen.cleanClonedRepo(destDir);
     };
     /**
      * if ( init git project )
@@ -59,12 +54,12 @@ var CommonGen = (function () {
      */
     CommonGen.prototype.generate = function () {
         if (this.config.repository.common) {
-            if (this.config.repository['firstTime']) {
-                return this.initFirstTimeCommonProject();
+            if (!GitGen_1.GitGen.commonProjectExists) {
+                this.createCommonProject();
             }
-            return Promise.resolve();
+            return this.addSubModule();
         }
-        return this.initWithoutSubModule();
+        this.initWithoutSubModule();
     };
     return CommonGen;
 }());