+++ CHANGELOG.md
@@ -1,10 +1,23 @@
-# 5.1.0
+# 0.5.4
 ### Features
-* Deploy
-### Bug Fixes
+* Deploy history
 
-# 5.2.0
+### Bug Fixes
+* updating docker-compose production in order to be testable in windows
+* creating master branch after initiating git repository on generated project
+* deploy
+
+# 0.5.3
+### Bug Fixes
+* package.json path
+
+# 0.5.2
 ### Features
 * Backup
+
 ### Bug Fixes
+* Deploy
+
+# 0.5.1
+### Features
 * Deploy